@@ -12,25 +12,24 @@
   "dependencies": {
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/data-grid": "^4.0.0-alpha.37",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.61",
-    "@mui/icons-material": "^5.15.21",
+    "@mui/icons-material": "^5.16.7",
+    "@mui/lab": "^5.0.0-alpha.173",
     "@mui/material": "^5.16.7",
+    "@mui/x-data-grid": "^7.12.1",
     "@stripe/react-stripe-js": "^2.7.3",
     "@stripe/stripe-js": "^4.1.0",
+    "@types/react": "^18.3.0",
     "axios": "^1.7.2",
     "chart.js": "^4.4.3",
     "country-state-city": "^3.2.1",
     "ecommerce": "file:..",
     "js-cookie": "^3.0.5",
     "overlay-navbar": "^1.2.3",
-    "react": "^18.2.0",
+    "react": "18.3.0",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "18.3.0",
     "react-helmet": "^6.1.0",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.2.1",
@@ -45,7 +44,6 @@
     "webfontloader": "^1.6.28"
   },
   "devDependencies": {
-    "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@vitejs/plugin-react": "^4.2.1",
     "eslint": "^8.57.0",
