@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import SideBar from '../SideBar/SideBar';
 import './DashBoard.css';
 import { Link } from 'react-router-dom';
-import Typography from '@material-ui/core/Typography';
+import Typography from "@mui/material/Typography";
 import {
   Chart as ChartJS,
   CategoryScale,
