@@ -2,10 +2,10 @@ import React, { useEffect, useState } from 'react';
 import './ProcessOrder.css';
 import { useDispatch, useSelector } from 'react-redux';
 import { Link, useNavigate, useParams } from 'react-router-dom';
-import { Button, Typography } from '@material-ui/core';
+import { Button, Typography } from '@mui/material';
 import { clearError, singleOrderDetail, updateOrder } from '../../actions/orderAction';
 import Loader from '../../Components/Loader/Loder';
-import AccountTreeIcon from '@material-ui/icons/AccountTree';
+import AccountTreeIcon from '@mui/icons-material/AccountTree';
 import { useAlert } from 'react-alert';
 import { UPDATE_ORDER_RESET } from '../../constants/orderConstants';
 import SideBar from '../SideBar/SideBar';
