@@ -1,7 +1,7 @@
 
 import React from "react";
 import "./Contact.css";
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material/Button";
 
 const Contact = () => {
   return (
